using browserify for some client deps
