+++ views/assets/c3.jade
@@ -1,0 +1,3 @@
+link(rel='stylesheet', href='//cdn.rawgit.com/masayuki0812/c3/master/c3.css')
+script(src='//cdn.jsdelivr.net/d3js/latest/d3.min.js')
+script(src='//cdn.rawgit.com/masayuki0812/c3/master/c3.js')
