+++ views/assets/calheatmap.jade
@@ -1,0 +1,2 @@
+link(rel="stylesheet", href='//cdn.rawgit.com/kamisama/cal-heatmap/master/cal-heatmap.css')
+script(src='//cdn.rawgit.com/kamisama/cal-heatmap/master/cal-heatmap.js')
