+++ views/match.jade
@@ -45,9 +45,4 @@ block content
         .addthis_sharing_toolbox.pull-right
 
 block footer_assets
-  link(rel="stylesheet", href='//cdn.jsdelivr.net/qtip2/latest/jquery.qtip.min.css')
-  script(src='//cdn.jsdelivr.net/momentjs/latest/moment.min.js')
   script(type="text/javascript", src="//s7.addthis.com/js/300/addthis_widget.js#pubid=ra-5452f1e51eac0fc5", async="async")
-  script(src='//cdn.jsdelivr.net/qtip2/latest/jquery.qtip.js')
-  script.
-    $('[title!=""]').qtip({style: "qtip-dark"});
