+++ views/match_chat.jade
@@ -30,5 +30,3 @@ append content
     h1 This match has no parsed data!
     
 append footer_assets
-  script
-    include ../public/js/yasp.js
