+++ views/match_details.jade
@@ -148,6 +148,3 @@ append content
         h1 This match has no parsed data!
 
 append footer_assets
-  script(src='//cdn.rawgit.com/adamwdraper/Numeral-js/master/numeral.js')
-  script
-    include ../public/js/yasp.js
