+++ views/match_index.jade
@@ -97,8 +97,3 @@ append content
       include ads/skyscraper
       
 append footer_assets
-  script(src='//cdn.rawgit.com/adamwdraper/Numeral-js/master/numeral.js')
-  script
-    include ../public/js/yasp.js
-  script
-    include ../public/js/map.js
