+++ views/matches.jade
@@ -1,11 +1,5 @@
 extend layout
 
-append head
-  script.
-    var parse_status = !{JSON.stringify(constants.parse_status)},
-        regions = !{JSON.stringify(constants.regions)},
-        modes = !{JSON.stringify(constants.modes)}
-    
 block content
   h3 Recent Matches
   table.table#table
@@ -13,6 +7,5 @@ block content
   
 block footer_assets
  include assets/datatables
- script(src='//cdn.jsdelivr.net/momentjs/latest/moment.min.js')
- script
-   include ../public/js/match_table.js
+ script.
+   matchTables()
