+++ views/player.jade
@@ -27,4 +27,3 @@ block content
              li(class= (route === key ? 'active' : ''))
                  a(href="/players/#{player.account_id}/" + (key == 'index' ? "" : key))= val.name
      block playercontent 
-
